moved message & user to main package
