+++ app/src/main/java/com/gpayinterns/chat/MessageRecyclerAdapter.java
@@ -17,7 +17,7 @@ import java.util.List;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
-import util.Message;
+import com.gpayinterns.chat.Message;
 
 public class MessageRecyclerAdapter extends RecyclerView.Adapter <MessageRecyclerAdapter.ViewHolder>
 {
